refactor(product): move forms + UI filters into dojo/product/ui/ [product Phase 3,4]
